+++ module/data/item/equipment.d.ts
@@ -1,0 +1,11 @@
+export {};
+
+declare module "./equipment.mjs" {
+	interface Equipment {
+		type: keyof typeof curseborne.config.equipmentTypes;
+		equipped: boolean;
+
+		armor: number;
+		enhancement: number;
+	}
+}